chore: move validateContractConfigs to config construction
