+++ apps/ensindexer/src/config/validations.ts
@@ -1,4 +1,8 @@
 import { ENSIndexerConfig } from "@/config/types";
+import { SELECTED_ENS_DEPLOYMENT } from "@/lib/globals";
+import { DatasourceName, ENSDeploymentChain } from "@ensnode/ens-deployments";
+import { PluginName } from "@ensnode/utils";
+import { Address, isAddress } from "viem";
 import { MergedPonderConfig } from "../../ponder.config";
 
 // Helper function to check rpc url exists for a given chain
@@ -81,4 +85,32 @@ export function validateConfig(config: ENSIndexerConfig, ponderConfig: MergedPon
   // most one network.
   ////////
   validateGlobalBlockrange(config, ponderConfig);
+}
+
+/**
+ * Validates runtime contract configuration.
+ *
+ * @param contracts - An array of contract configurations to validate
+ * @throws {Error} If any contract with an address field has an invalid address format
+ */
+export function validateContractConfigs(
+  pluginName: PluginName,
+  ensDeploymentChain: ENSDeploymentChain,
+) {
+  const contracts = SELECTED_ENS_DEPLOYMENT[DatasourceName.Root].contracts;
+
+  // invariant: `contracts` must provide valid addresses if a filter is not provided
+  //  (see packages/ens-deployments/src/ens-test-env.ts) for context
+  const hasAddresses = Object.values(contracts)
+    .filter((config) => "address" in config) // only ContractConfigs with `address` defined
+    .every((config) => isAddress(config.address as Address)); // must be a valid `Address`
+
+  if (!hasAddresses) {
+    throw new Error(
+      `The ENSDeployment '${ensDeploymentChain}' provided to the '${pluginName}' plugin does not define valid addresses. This occurs if the 'address' of any ContractConfig in the ENSDeployment is malformed (i.e. not an Address). This is only likely to occur if you are running the 'ens-test-env' ENSDeployment outside of the context of the ens-test-env tool (https://github.com/ensdomains/ens-test-env). If you are activating the ens-test-env plugin and receive this error, NEXT_PUBLIC_DEPLOYMENT_ADDRESSES or DEPLOYMENT_ADDRESSES is not available in the env or is malformed.
+
+Here are the contract configs we attempted to validate:
+${JSON.stringify(contracts)}`,
+    );
+  }
 }
